requirements.txt
Django==6.0.2
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
psycopg2-binary==2.9.11
django-environ==0.12.1
gunicorn==25.1.0